[#33] some cleanup of utilities code - more todo
